refactor: adopt components/ui/ for shared primitives
Moves the 9 components with 2+ consumers that are genuinely
generic, reusable UI primitives - not domain/feature composites -
into components/ui/: FileInput, HeadingText, Input, InputSelect,
Modal, PrimaryButton, PrimaryLinkButton, PrivacyPolicyModal,
Spinner.

Everything else in components/ (single-consumer sections, and
multi-consumer feature/layout composites like Companies, Profile,
Faq, TopBar) stays flat at the top level - route-local
_components/ colocation under app/ was considered but rejected in
favor of keeping all components under components/.
