+++ src/components/Companies/CompanyProfile/CompanyStatsSection/index.tsx
@@ -7,8 +7,8 @@ import { toast } from "react-toastify";
 import type { Stats } from "@/types/Stats";
 import { httpClient } from "@/lib/http/client";
 import { useMutation } from "@/hooks/useMutation";
+import PrimaryButton from "@/components/ui/PrimaryButton";
 import HistorySection from "@/components/HistorySection";
-import PrimaryButton from "@/components/PrimaryButton";
 import InterestSelector from "@/components/Profile/InterestSelector";
 import type { SavedStudentDto } from "@/application/dto/historyDto";
 import type { InterestDto } from "@/application/dto/interestDto";
