+++ src/components/Companies/CompanySocials/index.tsx
@@ -2,6 +2,7 @@
 
 import React from "react";
 
+import HeadingText from "@/components/ui/HeadingText";
 import {
   Facebook,
   Globe,
@@ -10,7 +11,6 @@ import {
   Twitter,
   Youtube,
 } from "@/components/ui/Icons";
-import HeadingText from "@/components/HeadingText";
 import SocialMediaCard from "@/components/SocialMediaCard";
 
 interface CompanySocialsProps {
