+++ src/components/DownloadButton/index.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 
 import { httpClient } from "@/lib/http/client";
 import { DownloadIcon } from "@/components/ui/Icons";
-import Spinner from "@/components/Spinner";
+import Spinner from "@/components/ui/Spinner";
 import { branding } from "@/edition/branding";
 
 import download from "downloadjs";
