+++ src/components/InfoText/index.tsx
@@ -3,8 +3,8 @@
 import { motion } from "framer-motion";
 
 import { QuestionCircle } from "@/components/ui/Icons";
+import PrimaryLinkButton from "@/components/ui/PrimaryLinkButton";
 import EventInfos from "@/components/EventInfos";
-import PrimaryLinkButton from "@/components/PrimaryLinkButton";
 
 interface InfoTextProps {
   days: number[];
