+++ src/components/Profile/ProfileSection/index.tsx
@@ -9,8 +9,8 @@ import { toast } from "react-toastify";
 import { ProfileData } from "@/types/ProfileData";
 import { httpClient } from "@/lib/http/client";
 import { useMutation } from "@/hooks/useMutation";
-import Modal from "@/components/Modal";
-import PrimaryButton from "@/components/PrimaryButton";
+import Modal from "@/components/ui/Modal";
+import PrimaryButton from "@/components/ui/PrimaryButton";
 import AvatarCropper from "@/components/Profile/AvatarCropper";
 import ImportCvSection from "@/components/Profile/ImportCvSection";
 import Input from "@/components/Profile/Input";
