+++ src/components/Schedule/index.tsx
@@ -4,7 +4,7 @@ import React, { useState } from "react";
 import useWindowSize from "@rooks/use-window-size";
 import { motion } from "framer-motion";
 
-import HeadingText from "@/components/HeadingText";
+import HeadingText from "@/components/ui/HeadingText";
 
 export interface ScheduleDay {
   hour: string;
