+++ src/components/SignUp/EmailStep/index.tsx
@@ -10,8 +10,8 @@ import {
 import { motion } from "framer-motion";
 
 import { StudentSignUpData } from "@/types/StudentSignUpData";
-import Input from "@/components/Input";
-import PrimaryButton from "@/components/PrimaryButton";
+import Input from "@/components/ui/Input";
+import PrimaryButton from "@/components/ui/PrimaryButton";
 import { isIsepEmail } from "@/utils/isepEmail";
 
 interface EmailStepProps {
