+++ src/components/SignUp/PasswordStep/index.tsx
@@ -11,8 +11,8 @@ import Image from "next/image";
 import { motion } from "framer-motion";
 
 import { StudentSignUpData } from "@/types/StudentSignUpData";
-import Input from "@/components/Input";
-import PrimaryButton from "@/components/PrimaryButton";
+import Input from "@/components/ui/Input";
+import PrimaryButton from "@/components/ui/PrimaryButton";
 
 interface PasswordStepProps {
   currentStep: number;
